+++ src/index.html
@@ -14,20 +14,19 @@
   </head>
   <body>
     <!-- ❗️❗️❗️ Add partials like so for every page -->
-    <include src="./partials/andrii-modal-film.html"></include>
-    <include src="./partials/header.html"></include>
+      <include src="./partials/modal-film.html"></include>
+      <include src="./partials/header.html"></include>
     <main>
       <include src="./partials/firebase-auth-modal.html"></include>
       <include src="./partials/gallery.html"></include>
       <include src="./partials/pagination.html"></include>
     </main>
-    <include src="./partials/footer.html"></include>
-    <include src="./partials/modal_team.html"></include>
-    <include src="./partials/top-button.html"></include>
+      <include src="./partials/footer.html"></include>
+      <include src="./partials/modal_team.html"></include>
+      <include src="./partials/top-button.html"></include>
 
     
 
     <script type="module" src="./index.js"></script>
-    <script src="./js/modal-film.js"></script>
-  </body>
+    </body>
 </html>
