+++ src/index.js
@@ -5,7 +5,7 @@ import './js/load-homepage';
 import './js/modal-team'; 
 import * as firebase from "./js/firebase-auth.js"
 import './js/top-button';
-import './js/andrii_modal_film';
+import './js/modal-film';
