--- src/js/pages-btn.js
@@ -1,1 +1,0 @@
-// ---- Создать класс PagesButton ------ //
