+++ src/partials/modal-film.html
@@ -1,0 +1,3 @@
+<div id="myModal" class="modal"></div>
+  
+
